assert()s só em modo DEBUG. Passos em unsigned long int.

+++ capsule.h
@@ -73,7 +73,7 @@ typedef struct SCapsule {
 	v3d    pos,     // posição inicial
 	       vel;     // velocidade
 
-	unsigned int steps;
+	unsigned long int steps;
 
         Ring* anel;     // aneis
